Add files via upload
For this project, I use real-world data for my experiments.  This data has a lot of missing values and noises. The goal of this project is the survival of patients with organ transplantation. I performed my work in 4 steps: first, I prepared the data by removing high missing value rows, select the features with expert article, normalize the data, second, impute the data with knn imputation method. Third, balance with label propagation, SMOTE and Kernel_SMOTE. Fourth, classification with seven methods. Finally, I obtained 105 models, which are the seven models selected by the seven classifiers as the best prediction model. In the next step, I prioritize the patients who have a positive result of transplantation prediction.